Work without `cryptography` as far as possible
Previously, an `ImportError` caused by a missing `cryptography`
library was only partially accounted for, which would have crashed the
execution module. Although `cryptography` is a core requirement for
Salt, there might be situations where it is not available (e.g. when run
via Salt-SSH). Don't cause an ugly crash then and provide functionality,
as far as possible.

Also, properly account for x509_v2 not being active in sign_certificate.
